all input external
