--- WINDOW_openMDAO/Input/ct_dtu10.dat
@@ -1,22 +1,0 @@
-4.0 0.923
-5.0 0.919
-6.0  0.904
-7.0  0.858
-8.0  0.814
-9.0  0.814
-10.0 0.814
-11.0 0.814
-12.0  0.577
-13.0  0.419
-14.0 0.323
-15.0 0.259
-16.0 0.211
-17.0 0.175
-18.0 0.148
-19.0 0.126
-20.0 0.109
-21.0 0.095
-22.0 0.084
-23.0 0.074
-24.0 0.066
-25.0 0.059
