--- WINDOW_openMDAO/Input/finance_parameters.py
@@ -1,4 +1,0 @@
-interest_rate = 0.075  # interest rate [-]
-# v = 0.015  # inflation rate [-]
-operational_lifetime = 25  # [years]
-management_percentage = 3.0  # [%]
