--- WINDOW_openMDAO/Input/power_dtu10.dat
@@ -1,22 +1,0 @@
-4.0 280200.0
-5.0 799100.0
-6.0 1532700.0
-7.0 2506100.0
-8.0 3730700.0
-9.0 5311800.0
-10.0 7286500.0
-11.0 9698300.0
-12.0 10639100.0
-13.0 10648500.0
-14.0 10639300.0
-15.0 10683700.0
-16.0 10642000.0
-17.0 10640000.0
-18.0 10639900.0
-19.0 10652800.0
-20.0 10646200.0
-21.0 10644000.0
-22.0 10641200.0
-23.0 10639500.0
-24.0 10643600.0
-25.0 10635700.0
