--- WINDOW_openMDAO/Input/turbine_parameters.py
@@ -1,26 +1,0 @@
-from numpy import sqrt
-
-turbine_rated_power = 10000000.0  # [W]
-cutin_wind_speed = 4.0 # [m/s]
-cutout_wind_speed = 25.0 # [m/s]
-
-rotor_radius = 95.4  # [m]
-purchase_price = 10000000.0  # [Euro]
-warranty_percentage = 15  # [%]
-generator_voltage = 4000.0  # [V]
-
-hub_height = 119.0 # [m]
-solidity_rotor = 0.0516  # [-] 'Generic' value, based on Peter Jamieson's book - Figure 2.5 - P.53
-cd_rotor_idle_vane = 0.4  # [-] 'Generic' value, very dependent on angle of attack and therefore the assumed rotor misalignment
-cd_nacelle = 1.2  # [-]
-front_area_nacelle = 14.0  # [m^2] 
-max_thrust = 1500000.0  # [N]
-yaw_to_hub_height = 5.01  # [m]
-mass = 589211.0  # [kg]
-mass_eccentricity = 1.9  # [m] - in x-direction, so negative when upwind of tower centre
-yaw_diameter = 5.5  # [m]
-wind_speed_at_max_thrust = 11.4  # [m/s]
-
-collection_voltage = 66000.0 # [V]
-turbine_rated_current = turbine_rated_power / (collection_voltage  * sqrt(3.0))  # A = Power / sqrt(3) / Voltage. 3 phase current per line.
-
