--- WINDOW_openMDAO/Input/weibull_windrose_12identical.dat
@@ -1,12 +1,0 @@
-0.0 8.15 2.11 8.333333333333333333
-30.0 8.15 2.11 8.333333333333333333
-60.0 8.15 2.11 8.333333333333333333
-90.0 8.15 2.11 8.333333333333333333
-120.0 8.15 2.11 8.333333333333333333
-150.0 8.15 2.11 8.333333333333333333
-180.0 8.15 2.11 8.333333333333333333
-210.0 8.15 2.11 8.333333333333333333
-240.0 8.15 2.11 8.333333333333333333
-270.0 8.15 2.11 8.333333333333333333
-300.0 8.15 2.11 8.333333333333333333
-330.0 8.15 2.11 8.333333333333333333
