+++ WINDOW_openMDAO/Turbine/Curves.py
@@ -1,6 +1,9 @@
 from WINDOW_openMDAO.src.api import AbsTurbine
 from aero_models import AeroLookup
 from WINDOW_openMDAO.input_params import cutin_wind_speed as cutin, cutout_wind_speed as cutout
+
+print cutin
+from WINDOW_openMDAO.input_params import power_curve_path, ct_curve_path
 import os
 
 
@@ -8,7 +11,7 @@ class Curves(AbsTurbine):
 
     def turbine_model(self, u):
 
-        table = AeroLookup(os.path.join(os.path.dirname(__file__), "../Input/power_dtu10.dat"))
+        table = AeroLookup(power_curve_path)
         if u < cutin:
             power = 0.0
         elif u <= cutout:
@@ -17,7 +20,7 @@ class Curves(AbsTurbine):
         else:
             power = 0.0
 
-        table = AeroLookup(os.path.join(os.path.dirname(__file__), "../Input/ct_dtu10.dat"))
+        table = AeroLookup(ct_curve_path)
         if u < cutin:
             ct = 0.0000001
         elif u <= cutout:
@@ -26,3 +29,4 @@ class Curves(AbsTurbine):
             ct = 0.0000001
 
         return ct, power
+
