+++ WINDOW_openMDAO/input_params.py
@@ -1,4 +1,5 @@
-from WINDOW_openMDAO.Input.farm_parameters import *
-from WINDOW_openMDAO.Input.turbine_parameters import *
-from WINDOW_openMDAO.Input.finance_parameters import *
-from WINDOW_openMDAO.Input.site_parameters import *
+from Input.farm_parameters import *
+from Input.turbine_parameters import *
+from Input.finance_parameters import *
+from Input.site_parameters import *
+
