+++ example/IEA_borssele_irregular.py
@@ -14,7 +14,7 @@ def print_nice(string, value):
     print header
 prob = Problem()
 # prob.model = WorkingGroup(direction_sampling_angle=10.0, windspeed_sampling_points=5, windrose_file='Input/weibull_windrose_12identical.dat', power_curve_file='Input/power_dtu10.dat', ct_curve_file='Input/ct_dtu10.dat')
-prob.model = WorkingGroup(direction_sampling_angle=1.0, windspeed_sampling_points=10)
+prob.model = WorkingGroup(direction_sampling_angle=30.0, windspeed_sampling_points=7)
 prob.setup()
 
 print_nice("Time after setup", clock())
