+++ example/Input/weibull_windrose_12unique.dat
@@ -1,0 +1,12 @@
+0.0 8.65 2.11 5.1
+30.0  8.86 2.05 4.3
+60.0 8.15 2.35 4.3
+90.0  9.98 2.55 6.6
+120.0  11.35 2.81 8.9
+150.0  10.96 2.74 6.5
+180.0  11.28 2.63 8.7
+210.0 11.50 2.40 11.5
+240.0 11.08 2.23 12.0
+270.0 10.94 2.28 11.1
+300.0 11.27 2.29 11.4
+330.0 10.55 2.28 9.6
